Seperate test for golden rules and clean param rules
106 english test passing
